+++ hw15-MessageSystem/src/main/java/ru/otus/springmvcwebapp/chat/ChatController.java
@@ -1,4 +1,4 @@
-package ru.otus.springmvcwebapp.hello;
+package ru.otus.springmvcwebapp.chat;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
