+++ hw15-MessageSystem/src/main/java/ru/otus/springmvcwebapp/chat/Message.java
@@ -1,4 +1,4 @@
-package ru.otus.springmvcwebapp.hello;
+package ru.otus.springmvcwebapp.chat;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
