+++ hw15-MessageSystem/src/main/java/ru/otus/springmvcwebapp/chat/MessageController.java
@@ -1,4 +1,4 @@
-package ru.otus.springmvcwebapp.hello;
+package ru.otus.springmvcwebapp.chat;
 
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.SendTo;
