+++ hw15-MessageSystem/src/main/java/ru/otus/springmvcwebapp/chat/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package ru.otus.springmvcwebapp.hello;
+package ru.otus.springmvcwebapp.chat;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
